migration: files section

+++ lua/sidebar-nvim/builtin/files.lua
@@ -5,8 +5,7 @@ local async = require("sidebar-nvim.lib.async")
 local Loclist = require("sidebar-nvim.lib.loclist")
 local utils = require("sidebar-nvim.utils")
 local has_devicons, devicons = pcall(require, "nvim-web-devicons")
-
-local api = async.api
+local logger = require("sidebar-nvim.logger")
 
 local files = Section:new({
     title = "Files",
@@ -30,8 +29,8 @@ local files = Section:new({
     trash_dir = vim.loop.os_homedir() .. "/.local/share/Trash/files/",
 
     reloaders = {
-        reloaders.autocmd({ "BufLeave" }, "*"),
-        reloaders.autocmd({ "ShellCmdPost" }, "term://*"),
+        reloaders.autocmd({ "BufLeave" }, "term://*"),
+        reloaders.autocmd({ "ShellCmdPost" }, "*"),
     },
 
     highlights = {
@@ -63,7 +62,6 @@ function files:scan_dir(directory)
         return
     end
 
-    local show_hidden = self.show_hidden
     local children = {}
     local children_directories = {}
     local children_files = {}
@@ -79,7 +77,7 @@ function files:scan_dir(directory)
         end
 
         if not ignored then
-            if show_hidden or filename:sub(1, 1) ~= "." then
+            if self.show_hidden or filename:sub(1, 1) ~= "." then
                 if filetype == "file" then
                     table.insert(children_files, {
                         name = filename,
@@ -116,52 +114,50 @@ end
 function files:build_loclist(directory, level)
     local items = {}
 
-    if directory.children then
-        for _, node in ipairs(directory.children) do
-            if node.type == "file" then
-                local icon = self:get_fileicon(node.name)
-                local selected = { text = "" }
+    for _, node in ipairs(directory.children or {}) do
+        if node.type == "file" then
+            local icon = self:get_fileicon(node.name)
+            local selected = { text = "" }
 
-                if yanked_files[node.path] then
-                    selected = { text = " *", hl = "SidebarNvimFilesYanked" }
-                elseif cut_files[node.path] then
-                    selected = { text = " *", hl = "SidebarNvimFilesCut" }
-                end
-
-                table.insert(
-                    items,
-                    LineBuilder:new(self:create_keymaps(node))
-                        :left(string.rep("  ", level) .. icon.text .. " ", icon.hl)
-                        :left(node.name)
-                        :left(selected.text, selected.hl)
-                )
-            elseif node.type == "directory" then
-                local icon
-                if self.open_directories[node.path] then
-                    icon = self.tree_icons["directory_open"]
-                else
-                    icon = self.tree_icons["directory_closed"]
-                end
-
-                local selected = { text = "" }
-
-                if self.yanked_files[node.path] then
-                    selected = { text = " *", hl = "SidebarNvimFilesYanked" }
-                elseif cut_files[node.path] then
-                    selected = { text = " *", hl = "SidebarNvimFilesCut" }
-                end
-
-                table.insert(
-                    items,
-                    LineBuilder:new(self:create_keymaps(node))
-                        :left(string.rep("  ", level) .. icon .. " " .. node.name, "SidebarNvimFilesDirectory")
-                        :left(selected.text, selected.hl)
-                )
+            if self.yanked_files[node.path] then
+                selected = { text = " *", hl = "SidebarNvimFilesYanked" }
+            elseif self.cut_files[node.path] then
+                selected = { text = " *", hl = "SidebarNvimFilesCut" }
             end
 
-            if node.type == "directory" and self.open_directories[node.path] then
-                vim.list_extend(items, self:build_loclist(node, level + 1))
+            table.insert(
+                items,
+                LineBuilder:new({ keymaps = self:create_keymaps(node) })
+                    :left(string.rep("  ", level) .. icon.text .. " ", icon.hl)
+                    :left(node.name)
+                    :left(selected.text, selected.hl)
+            )
+        elseif node.type == "directory" then
+            local icon
+            if self.open_directories[node.path] then
+                icon = self.tree_icons["directory_open"]
+            else
+                icon = self.tree_icons["directory_closed"]
             end
+
+            local selected = { text = "" }
+
+            if self.yanked_files[node.path] then
+                selected = { text = " *", hl = "SidebarNvimFilesYanked" }
+            elseif self.cut_files[node.path] then
+                selected = { text = " *", hl = "SidebarNvimFilesCut" }
+            end
+
+            table.insert(
+                items,
+                LineBuilder:new({ keymaps = self:create_keymaps(node) })
+                    :left(string.rep("  ", level) .. icon .. " " .. node.name, "SidebarNvimFilesDirectory")
+                    :left(selected.text, selected.hl)
+            )
+        end
+
+        if node.type == "directory" and self.open_directories[node.path] then
+            vim.list_extend(items, self:build_loclist(node, level + 1))
         end
     end
     return items
@@ -183,88 +179,113 @@ local function undo(group)
 end
 
 local function copy_file(src, dest, confirm_overwrite)
-    if confirm_overwrite and luv.fs_access(dest, "r") ~= false then
-        local overwrite = vim.fn.input('file "' .. dest .. '" already exists. Overwrite? y/n: ')
-
-        if overwrite ~= "y" then
-            return
+    local copy_cb = function()
+        local err = async.uv.fs_copyfile(src, dest)
+        if err ~= nil then
+            logger:error("error trying to copy file: " .. tostring(err), { section = "files", src = src, dest = dest })
         end
     end
 
-    luv.fs_copyfile(src, dest, function(err, _)
-        if err ~= nil then
-            vim.schedule(function()
-                utils.echo_warning(err)
-            end)
-        end
-    end)
-end
+    if confirm_overwrite and vim.loop.fs_access(dest, "r") then
+        async.ui.input({ prompt = 'file "' .. dest .. '" already exists. Overwrite? y/n: ' }, function(overwrite)
+            if overwrite ~= "y" then
+                return
+            end
 
-local function create_file(dest)
-    if luv.fs_access(dest, "r") ~= false then
-        vim.schedule(function()
-            utils.echo_warning('file "' .. dest .. '" already exists.')
+            copy_cb()
         end)
         return
     end
 
+    copy_cb()
+end
+
+local function create_file(dest)
+    if vim.loop.fs_access(dest, "r") then
+        logger:error(
+            'file "' .. dest .. '" already exists.',
+            { section = "files", operation = "create_file", dest = dest }
+        )
+        return
+    end
+
     local is_file = not dest:match("/$")
-    local parent_folders = vim.fn.fnamemodify(dest, ":h")
+    local parent_folders = async.fn.fnamemodify(dest, ":h")
 
     if not utils.file_exist(parent_folders) then
-        local success = vim.fn.mkdir(parent_folders, "p")
+        local success = async.fn.mkdir(parent_folders, "p")
         if not success then
             utils.echo_warning("Could not create directory " .. parent_folders)
         end
     end
 
     if is_file then
-        luv.fs_open(dest, "w", 420, function(err, file)
-            if err ~= nil then
-                vim.schedule(function()
-                    utils.echo_warning(err)
-                end)
-            else
-                luv.fs_close(file)
-            end
-        end)
+        local err, fd = async.uv.fs_open(dest, "w", 420)
+        if err ~= nil then
+            logger:error(
+                "error trying to create file: " .. tostring(err),
+                { section = "files", operation = "create_file+fs_open", dest = dest, err = err }
+            )
+            return
+        end
+
+        err = async.uv.fs_close(fd)
+        if err ~= nil then
+            logger:error(
+                "error trying to close file: " .. tostring(err),
+                { section = "files", operation = "create_file+fs_close", dest = dest, err = err }
+            )
+            return
+        end
     end
 end
 
 local function delete_file(src, trash, confirm_deletion)
-    if confirm_deletion then
-        local delete = vim.fn.input('delete file "' .. src .. '"? y/n: ')
-
-        if delete ~= "y" then
-            return
+    local delete_cb = function()
+        local err = async.uv.fs_rename(src, trash)
+        if err ~= nil then
+            logger:error(
+                "error trying to move file to trash: " .. tostring(err),
+                { section = "files", src = src, dest = trash }
+            )
         end
     end
 
-    luv.fs_rename(src, trash, function(err, _)
-        if err ~= nil then
-            vim.schedule(function()
-                utils.echo_warning(err)
-            end)
-        end
-    end)
+    if confirm_deletion then
+        async.ui.input({ prompt = 'delete file "' .. src .. '"? y/n: ' }, function(value)
+            if value ~= "y" then
+                return
+            end
+
+            delete_cb()
+        end)
+        return
+    end
+
+    delete_cb()
 end
 
 local function move_file(src, dest, confirm_overwrite)
-    if confirm_overwrite and luv.fs_access(dest, "r") ~= false then
-        local overwrite = vim.fn.input('file "' .. dest .. '" already exists. Overwrite? y/n: ')
-
-        if overwrite ~= "y" then
-            return
+    local move_cb = function()
+        local err = async.uv.fs_rename(src, dest)
+        if err ~= nil then
+            logger:error("error trying to move file: " .. tostring(err), { section = "files", src = src, dest = dest })
         end
     end
 
-    luv.fs_rename(src, dest, function(err, _)
-        if err ~= nil then
-            vim.schedule(function()
-                utils.echo_warning(err)
-            end)
-        end
-    end)
+    if confirm_overwrite and vim.loop.fs_access(dest, "r") ~= false then
+        async.ui.input({ prompt = 'file "' .. dest .. '" already exists. Overwrite? y/n: ' }, function(overwrite)
+            if overwrite ~= "y" then
+                return
+            end
+
+            move_cb()
+        end)
+
+        return
+    end
+
+    move_cb()
 end
 
 function files:create_keymaps(node)
@@ -364,33 +385,35 @@ function files:create_keymaps(node)
 
             self.open_directories[parent] = true
 
-            local name = vim.fn.input("file name: ")
+            async.ui.input({ prompt = "file name: " }, function(name)
+                if not name or string.len(vim.trim(name)) == 0 then
+                    return
+                end
 
-            if string.len(vim.trim(name)) == 0 then
-                return
-            end
+                local operation
 
-            local operation
+                operation = {
+                    success = true,
+                    exec = function()
+                        create_file(operation.dest)
+                    end,
+                    undo = function()
+                        delete_file(operation.dest, self.trash_dir .. name, true)
+                    end,
+                    src = nil,
+                    dest = parent .. "/" .. name,
+                }
 
-            operation = {
-                success = true,
-                exec = function()
-                    create_file(operation.dest)
-                end,
-                undo = function()
-                    delete_file(operation.dest, self.trash_dir .. name, true)
-                end,
-                src = nil,
-                dest = parent .. "/" .. name,
-            }
+                local group = { executed = false, operations = { operation } }
 
-            local group = { executed = false, operations = { operation } }
+                self.history.position = self.history.position + 1
+                self.history.groups = vim.list_slice(self.history.groups, 1, self.history.position)
+                self.history.groups[self.history.position] = group
 
-            self.history.position = self.history.position + 1
-            self.history.groups = vim.list_slice(self.history.groups, 1, self.history.position)
-            self.history.groups[self.history.position] = group
+                exec(group)
 
-            exec(group)
+                self:invalidate()
+            end)
         end,
         -- open current file
         ["e"] = function()
@@ -407,28 +430,35 @@ function files:create_keymaps(node)
         end,
         -- rename
         ["r"] = function()
-            local new_name = vim.fn.input('rename file "' .. node.name .. '" to: ')
-            local operation
+            async.ui.input({ prompt = 'rename file "' .. node.name .. '" to: ' }, function(new_name)
+                if not new_name then
+                    return
+                end
 
-            operation = {
-                exec = function()
-                    move_file(operation.src, operation.dest, true)
-                end,
-                undo = function()
-                    move_file(operation.dest, operation.src, true)
-                end,
-                src = node.path,
-                dest = node.parent .. "/" .. new_name,
-            }
+                local operation
 
-            local group = { executed = false, operations = { operation } }
+                operation = {
+                    exec = function()
+                        move_file(operation.src, operation.dest, true)
+                    end,
+                    undo = function()
+                        move_file(operation.dest, operation.src, true)
+                    end,
+                    src = node.path,
+                    dest = node.parent .. "/" .. new_name,
+                }
 
-            self.history.position = self.history.position + 1
-            self.history.groups = vim.list_slice(self.history.groups, 1, self.history.position)
-            self.history.groups[self.history.position] = group
+                local group = { executed = false, operations = { operation } }
 
-            exec(group)
+                self.history.position = self.history.position + 1
+                self.history.groups = vim.list_slice(self.history.groups, 1, self.history.position)
+                self.history.groups[self.history.position] = group
+
+                exec(group)
+            end)
         end,
+        -- TODO: undo and redo are not line-based, but rather section-based 🤔
+        -- so here they only work if the cursor is hovering a loclist line
         -- undo
         ["u"] = function(_)
             if self.history.position > 0 then
@@ -461,7 +491,7 @@ function files:create_keymaps(node)
 end
 
 function files:draw_content()
-    local cwd = api.fn.getcwd()
+    local cwd = async.fn.getcwd()
     local group_name = utils.shortest_path(cwd)
 
     self.open_directories[cwd] = true
